requirements.txt
fastapi==0.92.0
weaviate-client==3.23.2
python-multipart==0.0.6
gradio==3.42.0
